Test oha through the Cargo-built binary
